requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
httpx==0.25.1
openai==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
httptools==0.6.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
